Add pixel tracker to track opened emails.
The strategy is to embed an IMG as plain HTML directly in body of an email.
When user opens it, the IMG loads and sends a GET request with some metadata
to its server.

The server responds with a 1x1 pixel that is invisible to human eye. Hence,
the user isn't aware that the email was tracked.

The CLI tool prompts for user confirmation before sending email.

The test_emaillab.py validates data in config file. Validation uses the
packages re and validate_email > py3-validate-email.

Creations:
-tests/test_emaillab.py

Updates:
-.gitignore
-Makefile
-Pipfile*
-src/emaillab/config.py
-src/emaillab/emaillab.py
-src/emaillab/main.py